// src/lib/types.ts
export type PaymentMethod = 'Cash' | 'Card' | 'BankTransfer' | 'Other';

export interface PaymentBreakdown {
    method: PaymentMethod;
    amount: number;
}